cmd/gopherbot: use releases cache for an hour
The previous condition could never be true (unless the current time
went back by more than an hour), so the cached releases.major value
was never used. An HTTP request to the remote server was made on each
getMajorReleases call.

This change fixes the condition to use the cached releases.major value
within an hour of fetching it.

Updates golang/go#24899

Change-Id: I71da1b2503075db878b371606436130931dbcc0d
Reviewed-on: https://go-review.googlesource.com/c/147437
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>