cmd/gopherbot: use maintner client to fetch releases
Previously, the https://golang.org/dl/?mode=json endpoint was used
to figure out the supported and upcoming Go releases. That endpoint
lists versions that have binary release files uploaded.

CL 146137 added an RPC endpoint to the maintner server to list
supported versions. It uses tags and branches in Gerrit to compute
that information, which makes it a better source as it's closer to
the canonical source of truth for Go releases. We can use it here,
and it allows the parsing code to be simpler.

Change-Id: Ida0142f01aa42d5660fd28d4f985ff07914f239d
Reviewed-on: https://go-review.googlesource.com/c/147438
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>