cmd/gerritbot: add informational message upon initial upload of CL
In some cases (https://golang.org/cl/148477), the original author
of a change has not registered themselves with Gerrit. This causes
the change to assign Gobot as the owner, making it difficult to know
who originally authored the change. When initially creating the
Gerrit CL, upload some informational text that includes which GitHub
PR the CL corresponds to and who the author is according to git.

Fixes golang/go#28662

Change-Id: Icc8dad4e79d7270635b71b1e2cd24e0e1b8e5286
Reviewed-on: https://go-review.googlesource.com/c/148564
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>