Background: The current maintner was reporting incorrect hashtags for certain commits.
Bug: The previous implementation was only looking at the last commit
message in order to parse out the hashtags. However, the last commit
message was not always a tag action. This function of displaying
the correct hashtags was not at all aware of the state of hashtags.

Fix: The current implementation looks at all commits regarding an
issue and works like a state machine by parsing the parent commit
messages as well as the current one. The states can be added, removed
or no action at all. Once all messages are parsed the correct
tags are joined together and returned correctly.

Fixes golang/go#28318