dashboard: skip writing snapshot for windows/arm builder
Writing the snapshot to GCS is timing out for the Windows/ARM builder.
It tars about 500MB of data, resulting in about 175MB of data, which
it then sends to GCS. It takes about 4:30 to tar the workdir/go directory
with the built-in tar command, leaving only 30s to transfer 175MB
across the web, so it's not surprising that it times out.

Updates golang/go#26148

Change-Id: I7e70aeaf1ac6334c17fdca64ae9ab946c3d87d9a
Reviewed-on: https://go-review.googlesource.com/c/153518
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>