cmd/buildlet: ignore symlinks for now
As part of getting the x/build tests passing (trybots + post-submit),
ignore the symlinks that are in the x/build repo for now.

We can add support when we actually need them for something. But I
imagine any test that needs them can & does just create them as
needed.

Updates golang/go#14594

Change-Id: I6a0584d35c8d0fc91b3cdc8114b473df7f0268c3
Reviewed-on: https://go-review.googlesource.com/c/157439
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>