cmd/coordinator: add linux build tag to 3 more .go files
CL 157441 added linux build tag for linux-only tools.
In cmd/coordinator, that was done for almost all .go files,
but builders_test.go, reverse_test.go, and trybot_dev.go were missed.

That causes the package to fail to build when being tested
on non-linux environments:

	$ go test golang.org/x/build/...
	...
	# golang.org/x/build/cmd/coordinator [golang.org/x/build/cmd/coordinator.test]
	cmd/coordinator/builders_test.go:14:2: undefined: handleBuilders
	FAIL	golang.org/x/build/cmd/coordinator [build failed]
	...

This change fixes that by adding the linux build tag to the
remaining files.

Change-Id: I942a3eb5dd499a1acd6e338d7ad795b5ea5de95d
Reviewed-on: https://go-review.googlesource.com/c/157498
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>