Updated version.sh in regards to Issue #29929.
As mentioned in issue #29929 :
The version.sh file attempts to append -$USER-$(date -u +%Y-%m-%dT%H:%M:%SZ) to the VERSION environment variable when the working directory is dirty, or when the current commit doesn't match the remote commit. When the current branch doesn't have an upstream set (as can be the case if checking out a CL locally), git rev-list '@{upstream}..HEAD' exits with a non-zero exit code.
As a result, the VERSION variable is incorrectly set.
So, it can be corrected by a small check using "git ls-remote" command. It ensures that if an upstream doesn't exist USER and DATE are appended to VERSION, else it continues to next check " git rev-list '@{upstream}..HEAD' ".