cmd/updatecontrib: overhaul output to be useful
The updatecontrib command came to be in CL 157238 by reducing the scope
of a previous updateac command. The updateac command was modified to
fit the needs of updatecontrib. The modification was aiming to be
minimal in nature, it focused on removing CLA checking code and removing
the functionality that updated the AUTHORS file.

Most of the output that the new updatecontrib command produced was only
relevant to the original updateac program, and wasn't very helpful for
the purposes of updatecontrib.

At that time, I did not have a complete understanding of the output,
so I chose to make minimal modifications that I understood. By now,
I've gained that understanding and was able to overhaul the output of
updatecontrib to make it useful, as well as refactor the code to make
it simpler and more readable. There is no change in behavior.

The output is now separated into 3 categories:

• actions  - list of actions taken during the run
• warnings - list of situations where a valid-looking name couldn't be
             determined, causing the git commit author to not be added
             to the CONTRIBUTORS file
• errors   - list of unexpected errors (e.g., a network error talking
             to GitHub API)

Warnings are a normal outcome, since there are many git commit
authors that don't have a name set in any of the sources that
updatecontrib currently checks.

Errors are not expected, and are no longer treated the same way as
harmless warnings, which made output harder to read.

Factor out code that queries GitHub API into a new github.go file.

Add a new nameFix entry to use a normalized name capitalization.

Change-Id: I433ebc5afd8986709b5f323358efc93e3995b2b8
Reviewed-on: https://go-review.googlesource.com/c/160277
Reviewed-by: Andrew Bonventre <andybons@golang.org>