maintner,devapp: add ability to filter changes on reviews page
This change adds a filter input field to the /reviews page,
allowing people to refine what is shown. When updating the reviews
data, it collects search terms that are then rendered by the
template in data- attributes of the corresponding rows.

Filtering then does a match based on whether the row's
data attribute contains the terms entered (with the exception
being when one would like to exclude a term using a minus
sign (-) prefix.

A how-to is added to the UI as well, along with some aesthetic
changes to the rows.

Future work would involve including everything from a change
such as the commit message, hashtags, etc.

Also removes a superfluous log from maintner.

Change-Id: Ia6cbafcd08442dee1b48244725c09ea3575e69df
Reviewed-on: https://go-review.googlesource.com/c/163218
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>