cmd/coordinator: support building and testing on macOS
It's convenient to be able to build and run tests for cmd/coordinator
on macOS, in addition to Linux. Add the build tags and make appropriate
changes to the README to document the level of support.

I've tested running locally on macOS and the limited basic functionality
seems to work okay. If there's something specific that doesn't work on
macOS but works on Linux, we can update the README accordingly.

Start using go run instead of go install to combine two commands into
one. If someone prefers to use go install and run the installed binary,
they can infer the instructions to do so from the go run command.

Change-Id: I823b21a8c790c77045638b53c204ce10b60f40b2
Reviewed-on: https://go-review.googlesource.com/c/build/+/168137
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>