env/linux-x86-stretch: add netbase
Turns out I broke the linux-amd64 builder after CL 170282 because it
hadn't been deployed in a long time and there must've been some
refactoring in the Dockerfile meanwhile that didn't quite work.

Perhaps the --no-install-recommends was a recent addition and we were
previously getting the "netbase" package implicitly.

In any case, we need netbase to get /etc/services.

Deployed this & builds work again.

Change-Id: I9564d40a6bfe6f2606dafab10d66ffb914fcce18
Reviewed-on: https://go-review.googlesource.com/c/build/+/170283
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>