env/plan9: add amd64 support
This change adds support to generate amd64 Plan 9
image. Consequently the "plan9-386" directory was
renamed "plan9". The script handles either the
"386" or the "amd64" parameter to build the image.

This change also adds the possibility to chose to
build the image for QEMU or GCE by passing either
the "qemu" or the "gce" parameter.

Change-Id: Ic5aced903e0d1e7baa4ae7df2167c4d54cd6a66b
Reviewed-on: https://go-review.googlesource.com/c/build/+/4904
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>