cmd/buildlet, cmd/coordinator: trim space from reverse buildlets' gobuildkey file contents
Ignore whitespace before/after buildkeys in ~/.gobuildkey. It worked
in the past because this was only sent in HTTP headers where net/http
cleaned it. But once we also started sending it in HTTP Basic Auth
headers (where it's base64'ed), the newlines then made it to the
server where it didn't match.

So also clean it in the server for old clients.

Fixes golang/go#30749

Change-Id: I2d9b56909a0f616da670f4e41e63701a0872dff2
Reviewed-on: https://go-review.googlesource.com/c/build/+/170621
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>