dashboard: disable more builders
Make exp off by default, as it barely passes anywhere. Lock in the
places where it does pass.

Also disable x/net master combined with Go 1.11 on FreeBSD 32-bit. It
doesn't work on our builders and it'll never be fixed (we're not going
to make changes to Go 1.11) and there's resistance to skipping those
tests (my CL 170777), so I'm just turning off the builder. That's not
ideal, but it's better than people getting used to red on the
dashboard.

Fixes golang/go#31221

Change-Id: Id0228e4b0ee73715a0c2da84f68dcfc989a48aaf
Reviewed-on: https://go-review.googlesource.com/c/build/+/170782
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>