internal/gophers: document Person.Googler and Bot fields
The Googler field is determined via heuristics and represents whether
the person is known to be at Google now or has been in the past. It
can be used for gathering statistics about contributions that are
internal or external, but not for security related checks.

Make this more clear by documenting the field.

Change-Id: I966d3fa7a107c2db7b6898084db22a1d064c87cc
Reviewed-on: https://go-review.googlesource.com/c/build/+/175198
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>