cmd/coordinator: add missing Sprintf arguments in gitMirrorErrors
This is a followup to CL 179877 that adds missing arguments to
a fmt.Sprintf call.

Also change the if condition to ensure the regexp had exactly 3
matches as expected. That makes it more clear there won't be
a panic when accessing matches.

Change-Id: Ibf57d87565c8297cd68dd9a63b4dc709bce613de
Reviewed-on: https://go-review.googlesource.com/c/build/+/182017
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Emmanuel Odeke <emm.odeke@gmail.com>