dashboard: upsize machine type for windows-amd64-{longtest,race} builders
Start using n1-highcpu-8 machine type instead of n1-highcpu-4
for windows-amd64-longtest and windows-amd64-race builders.
Continue to use the existing n1-highcpu-4 machine type for all
other windows/386 and windows/amd64 builders for now.

The primary motivation of this change is to increase amount of
RAM available on the longtest builder from 3.6 GB to 7.2 GB,
because it is consistenly failing due to being out of memory.

Also fix minor typos, and increase consistency in comments and notes.

Updates golang/go#33951

Change-Id: Iac5da67fa584059842850a9c8deda98c242741f2
Reviewed-on: https://go-review.googlesource.com/c/build/+/208500
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>
Reviewed-by: Alexander Rakoczy <alex@golang.org>