cmd/coordinator: use time.UTC explicitly in tests
Also use the name “now” for both the Created and Expires fields on the
remoteBuildlet message, to avoid unnecessary (and mysterious) skew if
the call happens to span a second boundary.

This fixes TestHandleBuildetCreate_Stream and
TestHandleBuildetCreate_PreStream when run on a machine with a non-UTC
local time zone.

Change-Id: I507185c758bc4fc8b1c45e181a278816ee77a337
Reviewed-on: https://go-review.googlesource.com/c/build/+/208659
Run-TryBot: Bryan C. Mills <bcmills@google.com>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>