devapp: don't put release blocking issues in performance section
The performance section is intended to be lower priority.
If something is both a release blocker and a performance issue,
the release blocker tag takes precedence.

We already sort the issues correctly. But the code that decides where
the "performance" label goes is incorrect when there is an issue
which is both release blocker and performance.

Change-Id: I26c81ac380242b754862db7868d2dde8cdce00ed
Reviewed-on: https://go-review.googlesource.com/c/build/+/209677
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>