cmd/gerritbot: use the new repos package for repo policy
The only behavior change is that "gofrontend" is now included, which
is arguably a bug fix (an example of our various lists of repos
getting out of sync). At least, I don't recall any discussion of
omitting that repo. Ian +1'd this CL, so keeping them on.

If we want to exclude a repo from gerritbot in the future we'll need
to add some policy fields/methods to the repos package.

Updates golang/go#36047

Change-Id: I877cd3b6c292ac888863523e05799208e247499f
Reviewed-on: https://go-review.googlesource.com/c/build/+/210741
Reviewed-by: Alexander Rakoczy <alex@golang.org>