maintner: simplify TailNetworkMutationSource API
Document the property that MutationStreamEvent.Mutation is non-nil if
MutationStreamEvent.Err is nil. This is consistent with documentation
inside the MutationStreamEvent struct, and simplifies API for callers.
They can use the Mutation field directly after checking the Err field.

Use a generic temporary directory name in TailNetworkMutationSource,
since it can be used by anything, not just the maintwatch command.

Use the new godata.Server constant in maintwatch.

Change-Id: I39dd6918b5f21173a8f47a7359f489cab511295e
Reviewed-on: https://go-review.googlesource.com/c/build/+/221637
Reviewed-by: Alexander Rakoczy <alex@golang.org>