dashboard: add linux-amd64-staticlockranking builder
This is a new linux/amd64 GCE builder for the purpose of
testing the GOEXPERIMENT=staticlockranking configuration
which is landing to master branch soon.

It runs only on the main Go repository master branch and
release branches 1.15 and newer. No golang.org/x repos.
Not included in trybots by default, but can be requested
via SlowBots by saying TRY=linux-amd64-staticlockranking.

Also expand documentation of BuildConfig.buildsRepo field
to make it more clear how to use it. A common mistake for
custom buildsRepo functions that are added is to create a
custom policy with a minor adjustment and forgetting to
preserve the rest of the default build repo policy.

Fixes golang/go#37937

Change-Id: Ic56f82e9ede30cb6fe8f6cc7d0282d6a0a12c1b8
Reviewed-on: https://go-review.googlesource.com/c/build/+/224078
Reviewed-by: Robert Findley <rfindley@google.com>
Reviewed-by: Alexander Rakoczy <alex@golang.org>
Reviewed-by: Carlos Amedee <carlos@golang.org>
Run-TryBot: Alexander Rakoczy <alex@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>