internal/spanlog: copies spanlog into the internal directory
This is the first in a series of patches which include an
effort to refactor how the cmd/coordinator is structured. In order
to move buildlet pools into their own package, it would facilitate
the move to have spanlog in it's own internal package. This CL will
be followed by one which removes cmd/coordinator/spanlog and
changes the path of the package for all the callers.

Updates golang/go#36841

Change-Id: Ic310a38f807a3099815e0bdb988edc42b2ca5c85
Reviewed-on: https://go-review.googlesource.com/c/build/+/226843
Run-TryBot: Carlos Amedee <carlos@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Alexander Rakoczy <alex@golang.org>